Add files via upload
Changed the turn method to return what card a player played last if the player played a card, and to return spaces if the player didn't play anything. The player turn method now takes the last card played as an argument so that the card suit and number can be tracked for special rule purposes. Also added some more descriptive comments to methods.